--- twister2/common/src/java/edu/iu/dsc/tws/common/client/IClient.java
@@ -1,30 +1,0 @@
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-package edu.iu.dsc.tws.common.client;
-
-/**
- * An instance of this class will be executed in the submitting client,
- * if it is specified in Twister2Job
- */
-public interface IClient {
-
-  /**
-   * After the job is submitted,
-   * an instance of this method will be executed in the submitting client
-   *
-   * Implementing Client program can coomunicate with the JobMaster and workers
-   * with the provided  IClientController
-   *
-   * @param clientController
-   */
-  void execute(IClientController clientController);
-}
